requirements.txt
streamlit
pandas
deep-translator
qrcode
Pillow